Finish first prototype of reduce_deep over apps
This iteration confines everything into `reduce_deep` and is about to
be fragmented for use in binding functions (e.g. the handler for GHRS).
This prototype is still crude in its treatment of the stack, using fairly
inefficient update and query routines to look for variables and change the
symbol table in the appropriate frame. It does resolve lazily, only looking
for constructors for the sake of binding vars from within data.