Refactor stack
newtypes needed for Sym and SymTable because of mutual recusion. Also,
StackBranch went from being a sliver of a tree to a "true" branch, of
pointers between nodes. In the pure code, this stack will mirror the
stack of traversal which is why it can be just... well, a stack.