Begin removing explicit memoization scheme
I've proved that implicit memoization (i.e. matching thunks) won't work
for navigating the AST, which is a cyclic graph. To avoid cycles, we
need state, which changes the thunk and prevents auto memoization.

There's much more about this scattered around sandboxes (2019-07-13-0,
2019-07-10-0), in questions (SO
<https://stackoverflow.com/questions/57022334>,
<https://stackoverflow.com/questions/57000994>) and in an upcoming
blog post that I'm pretty sure will live at
<//lam.io/writing/cycle-memo>.

Explicit memoization is messy though, because there's a lot more
possible disagreement between the real result of `reduce_deep` and the
memoized result we're lugging around -- we need to worry about if the
memoized result is NF, which isn't enforced by the typechecker. Instead,
I'm opting to use the least efficient implementation of doing the least
amount of work possible on the StackBranch and calling `reduce_deep` at
all pattern matches. This is not as bad as it sounds, because intuitively
most values in programs are computed from a function call (which is unique
over its arguments) and used very locally; very rarely should values cross
many scope lines. Still, it's suboptimal, but wayyy easier to reason
about. After the first tests stabilize, I'll start to think about ways to
implement the explicit memoization.