Migrate SymTable & Sym -> Map & HsExpr
We can't use a custom data type in the return position (for partial
functions) and accept HsExpr in the argument, because partial functions
can be arguments to other functions, and we need to make sure that they
have the appropriate argument bindings but they can't be custom objects.
This is step 1: unwrap all the `Sym*` wrappers to their base data
counterparts on a type and constructor level.