Partial function: Sym's table -> nested HsApp
The only way to provide partial functions with arguments without the
custom data object (i.e. entirely with GHC expressions) is to tack the
arguments on explicitly with `HsApp` to the function. Then later the
`HsApp` pat match will re-decompose the arguments along with new ones
from the site of application, and maybe enter the function body.