Fix result duplication from overzealous `everything`
More specifically, looking for `GRHS` expressions to recurse into used
to include function bindings in `where` clauses (i.e. basically every
binding), as well as those in the guards of MatchGroups (which I didn't
know before except for the weirdness I saw in the `e` binding in the
`A.hs` test case).

After a lucky debug session, I discovered that:

1. With the new version of the `D.hs`, results were duplicated more often
the more bindings I added to the `where` clause;
2. When there was a binding introduced in a guard in one of the bindings,
some of the results wouldn't resolve the variable;
3. The `GRHS` returned one of the arguments in the `where` clause on top of
the variables/functions that were bound there;
4. It happened that the example I chose resolves the second instance very
soon after the first, so we see exactly what function the duplicate
expression comes from; in this case `grhs_expr`.

Thanks to this, I jus made `grhs_expr` more specific. Hooplah. There
went a few hours of debugging, but that would've been super confusing
later.