Extend all analysis methods with list of writes
In general, we need to know what normal forms are incident on pipes
because the writes are otherwise lost during `>>`, `<$`, `$>`, `void`
or other information-destroying commands. Therefore, we have the
`ReduceSyms` and `PatMatchSyms` datatypes that include maps from
something distinguishing pipe declarations (for now, SrcSpan) to
the values that are incident on them.

While we can separate statements in `reduce_deep` that won't introduce
any new writes based on types alone, it's just inelegant to separate
the function into a pure and impure instance. It's b/c the difference in
return type for pat matching and reducing differ in binding vars vs.
not respectively, and also pure vs. impure. However, pat_match is
scarcely different between pure vs. impure, yet we would have to rewrite
all the places that it draws values from `reduce_deep` vs. `reduce_deep_pure`
depending on call site. This is too annoying, so we'll just assume it's
all impure and go from there. Monoid helps for combining sub-results.

Tested against regressions to the target scripts A-D.