Add Located to symbols, begin handling `do`
Pipes are uniquely identified by the location of creation (and not even
just where they're instantiated either, because IOs can be re-run in
many locations) so we need to change Sym from HsExpr to LHsExpr.
Starting to think about handling the IO interface of IORefs here, where
to my surprise, the best place to handle them is not in `reduce_deep`,
but exclusively in a level above after calling `reduce_deep`.