Update Stack type in preparation for marking threads
Enriching the Stack type so that the places where threads are forked
are recorded to annotate the pipe writes. Under the restricted feature-
set where we don't consider recursive forks (e.g.
`mapM (const forkIO void) [0..]`), the stack branch leading up to a
fork is all the identifying information we need to uniquely identify
a thread.

This design competed against changing the `SrcSpan` in `StackBranch`
to an `LHsExpr`, but forking is the only use case that could need this
with a very specific form (i.e. always `HsApp` of `forkIO` and some `IO`
-returning thing) so being overt is more expressive in this case.