Distinguish BindStmts in `do` blocks from guards
The pattern binding `pat <- stmt` appears in both guards (e.g. function
guards, case statement guards, multi-if, etc.) and in do blocks. They
have very different semantics: the former matches the pattern to the
result on the right directly, where the latter matches the pattern to
the argument coming from the `>>=` refactor from the righthand side.
For transparent monads like `IO`, this means that while the former
pattern-matches against the value itself, the latter matches against the
"unwrapped" result.

This actually won't be put into practice yet and is just future-proofing
for now, because I'll be handling transparent monads... well,
transparently, so they'll be unwrapped internally in rahse, and we'll search
for the monad operations in the other non-magical monads.