Flag consumer values and associate instance bindings
This is requiring us to save extra state to the Sym, as expected from
the beginning (I needed to know if a value had been through the input
of a `Consumer`), but this change is causing problems: now, the `terminal`
can't both preserve the extra information _and_ compose them into an
`HsApp`, which takes `LHsExpr Id` specifically. This comes with a fairly
radical change to the structure of some of the code, because we had also
been chugging along hard with the assuming that `Sym === LHsExpr Id`.
In the end, between switching out of LHsExpr Id and special-casing the
result coming from `reduce_deep`, the latter is definitely more feasible
as far fewer things use it.

Instance bindings are also its own can of worms. Just see:
<https://stackoverflow.com/questions/58037699/ghc-api-finding-implementations-of-instance-functions>