Complete Sym/SymApp refactor & GHC 8.2.6 migration
Lots of structural errors due to breaking changes in datatype defs in
GHC were patched, and then the reduce routine was fully refactored to
operate over a richer Sym (enhanced with stack info and is_consumed).

There are plenty of new TODOs to verify if the stack locations of syms
are accurate/useful. There are many ways these rules can be written,
and there is one specific class that will fit this analysis that I need
come up with. There is also a list of things that should be todos: all
the instances of (<>) on Map instances that might want to be `unionWith (++)`
instead. At least `StackBranch` and `PatMatchSyms` use `unionWith (++)`.

Rerunning some of the examples (very superficially) has revealed some
regressions (very unsurprisingly), so step 0 might be to get these back
to stable.