Begin generalizing grhs_binds
Note that `do` and guard BindStmts have completely different semantics,
but it's not possible to tell which they represent based only on the
structure of the AST datatype. Instead, you have to look at the context
of the binding (i.e. the nearest parent expression) to know what kind
you're dealing with: an explicit pattern match (e.g. Just a <- fromMaybe...)
or a `do` binding (e.g. threadId <- forkIO ...).

This commit begins work to do that, although under initial assumptions
on the analysis performed by the library, this is not going to be
necessary: magic `do` bindings will in fact be pattern matched in the same
way as the patterns in guard statements. For the moment, this doesn't
typecheck either because of the many responsibilities it's balancing
that haven't been ironed out fully yet (e.g. tracking if we're in a `do`
vs. tracking other statements we shouldn't enter).