Patch WildPat match by refactoring pat_match
Many parts of the pattern-matching pipeline relied on adding matches
to `pms_syms` to signal success of the pattern match. WildPat is the
glaring exception to this, and in general justifies improving the
otherwise shaky check of pms_syms emptiness, so we've moved to a much
stronger `Maybe` signal. The notable exception is `bind_to_table`, which
for convenience and some degree of semantics unwraps its Maybe to an
empty `PatMatchSyms`.