Disintegrate Sym and move stack to SymApp
While all of the auxiliary data about symbols was starting to collect
in SymApp, I had no motivation to move the `stack_loc` out of Sym
so it remained a datatype. No motivation until now, that is.

`pat_match` and the resolution of variables in the stack were always
flawed. The former didn't take into account variables referring to each
other (e.g. `letrec` behavior). The latter didn't take into account the
possibility that the same function would be called more than once in a
stack from different locations in the code (e.g. the `$` operator).
The latter serves the former, because the pat_match now needs to
look at the syms that come out and see if they match the ones it's
currently evaluating. This is only correct if the stack matches.

Furthermore, this prepares the stage for hard and fast rules about the
propagation of auxiliary data. In this case, it's stack locations of
pipes, writes and consumers.