Begin SymApp stack audit
SymApp's stack is propagated in heterogeneous ways: function applications
update the stack by behavioral definition, but building a datatype, while
also an "application", distributes the stack across the records, rather
than, for example, inheriting the stack from the pattern matching context.

The situations are sometimes ambiguous, and sometimes in subtle ways (e.g.
in `reduce_deep`'s `HsVar` case, many stacks could be in scope at a given
time) so this is the beginning of a comprehensive scan of the places where
SymApps are updated, and the SymApp stacks are updated (or not).