Redo SymApp localization, remodel pattern matching
To the first point, full stacks are now used instead of StackKeys in
almost all places identifying syms uniquely. This is to hone the
cycle breaking and thread identification. The former was started in an
earlier commit because functions will be called multiple times but with
different arguments (e.g. `$`) and yet are not dangerous to the analysis.
This commit just moves all the functions to perform stack comparisons
(except pipe resolution, which is coming down the pipeline (pun)) and
follows rules about stack propagation more rigorously, with the central
role it will now be playing in correctness and halting. In particular,
I'm starting to chase down consistency to enforce rules in hold
resolution about the progression of stacks as holds are unravelled.
If the rules are inconsistent, then stacks made further along the
analysis might actually miss information and create very confusing
sensitivity problems at the very end.

To the second point, also as mentioned in 4bfc910, `letrec`-style mutual
references in binding clauses weren't being heeded before, and they are
a rather tricky case, being the only place where variables can refer to
each other. (Not even in argument pattern matching can these kinds of
references be created at the pattern binding site.) In addition, the
there needs to be an arbitrary number of reduction cycles, because
symbols that refer to each other need to wait for them to be reduced to
something outside of their mutual cycles (if any exist), pattern-matched,
then reduced again, and so on, until nothing new is substituted anymore.

Tests seem to be working nicely, `I.hs` was designed for these changes.
Many of the initial conditions are really ugly at the moment just to
get the ball rolling, and need to be refactored stat.