Make Stack equality based on equality of locations
This is to simplify and quicken the analysis instead of comparing whole
stacks and their inner branches (of stacks within SymApps). Part of the
deal is that, for the cycle detection to continue working properly at
the HsLam application site, all sa_args must be in normal form the moment
they're created. To be honest, this might've been a general necessity
prior anyways, and we just happened to solve two problems with the
guise that one was necessary to enable the other to be solved.

PS SectionR, if not already difficult enough, just got more bothersome.
Normally, we flip the arguments on the NF HsLam in the stack table.
However, the reference in the stack table could be a hold, so we can't
unless the application is also saturated in which case we can adjust the
order of tthe arguments in the SymApp. So now it fatals (a bit harsh
maybe) if the SectionR is totally partial.