Pattern match over Binds, and simplify `reduce` because of it
It turns out that Binds takes care of what used to be "Holds": pattern
matches that needed to be deferred while waiting for values to flow
through pipes. `reduce` is now greatly simplified into just re-matching
all the binds in the stacks of all the symbols in the state, and
re-running reduce_deep. So far, it's been able to flow nested writes
in the same binding scope.

Worth noting that Write locations are still wrong in this commit,
because they involve a `StackKey` comparison, but these don't contain
the location of the pipe itself, just the last function call. Therefore,
writes to any pipes in the same call are now matching, which is maaaybe
not what we want.