Apply `sub` to `Writes` in pat_match
Writes are quite unique: they are the only other category of expression
outside of binds and direct rhs / LastStmt for pure and impure contexts
resp. We apply `sub` to the arguments and lefthand sides of syms, then
flow the bindings via `reduce_deep`, but we never touch Writes. This
turned out to be a mistake, because then some Writes that bind to local
symbols would fail to find them when it came time to resolve them. This
was happening to the test file `H.hs` where a tuple `(v :: MVar Int, 42)`
was written to another MVar, then a read on `v` was attempted. It didn't
have the context to pull out the `newEmptyMVar` so the reduction stopped
at `v`.

That particular case still fails because of an unrelated problem with
treating tuples and lists as terminal symbols. That's the subject of
the next commit.