Special-case tuples and lists as symbols
As terminal symbols, tuple and list literals were causing problems with
substitutions and reduction, messing up a whole host of arguments like
"all pats coming out of pat_match are NF" and "all arguments are NF" and
"we've finished substituting all the symbols by this point", etc.

I felt that it is cleaner to special-case them explicitly rather than
convert them into vanilla SymApp with a special token at the front, but
that is a little harder to enforce and I think the semantics aren't quite
as clear.