Sub writes in pat_match iterant
The `J.hs` test case isn't resolving the `bar x` call because there is
no binding for `bar` whenever it gets resolved. Digging deeper, it turns
out that it's because the write is logged in `pat_match_one`, and writes
aren't being substituted, so when it's resolved at a later iteration,
`bar` is no longer available.

There are the beginnings of code to resolve writes, but it was in
retrospect just a stab at the symptom but not the cause: it shows off how
many writes are incident on that iterant: from the [Bind] arg, from
pattern matching, from substition on the returned symbols, and now
substitution on the writes.

Next commits will be attempts to pare this down and enforce rules on what
is reduced where.