Make `pat_match` generic, and the single point for deep reduction
Previously, the degree of reduction of SymApps in the application were
wildly inconsistent; at best I could just claim that "most" of them were
as normal-form as they could be for where they were coming from.
The codebase becomes immensely cleaner with this modification: grhs_binds
, sub and pat_match_\* all *try* to match symbols as needed, but will
only preserve the "normalness" of the input form.

Then, somewhat orthogonally but in the same direction, pat_match is now
generic in that it tries to `sub` all SymApps that aren't in the stack
and use the symbols from substitution + reduction (or lack thereof) as
the predicate for continued iteration or stopping.

Finally, there is a general problem with how bindings are given to
symbols that are themselves pattern-matched to produce those bindings.
This is necessary because by the current designed,  `sub` can't do all
the work of beta-reducing the current stack frame. "Holds" as they used
to be called -- calls that are dependent on a value from a pipe -- are
the responsibility of `reduce`, so `pat_match` would need to put the
binds in all symbols that it just expanded, so when the `reduce` write
resolver churns, it has the symbols to refer to.

This is something of a shortcoming of `Write`s I believe, because the
write won't be logged if the first argument isn't a raw pipe, like
"newEmptyMVar". This is so that we can maintain a lookup table, but
it seems now to be more useful to expand the table into a flat list and
have more expensive queries for the timebeing.
This is still a work in progress.