Allow Write objects to have unreduced pipes
Initially, when we had a lookup table for writes, we needed the reference
to be the correct, final pipe location. However, this was too strict,
because often pipes are referenced from the same binding scope, so
when they're reduced (in the `pat_match` stages notably), they aren't
bound because the pipe won't be resolved in general until the _end_ of
`pat_match`.

The fact that it's a many-many relationship between values and pipes is
a bit of a kludge to support the `everythingWithContextBut` in `pat_match`
which expands SymApps in general, so to make an isomorphism, the input
needs to itself be a list (in this case the pipe key).