Return all writes from `reduce` iterant
Making `reduce` truly `iterate` the `iterant` allows us to extract all
steps of the reduction, in particular collecting all the writes (since
only new writes are recorded in each iteration).

Main is the test ground for the violation test, which seems to be
straightforward enough: collect all the writes to true pipes that we
can find, see what threads they write from and what consumer they
come from (if any), then flag collisions.

We are also able to simplify the `iterant` type to
`ReduceSyms -> ReduceSyms` which is more expressive of the operation
that is going on.