Search for matching instance functions with soft type equality
Built-in Id equality acts on Unique. However, to match an instance
function, we have to use types because the instance isn't known until
the parameterization of the context of a given instance function call
is complete. The types are compared by "concreteness": instance methods
can only differ from their typeclass signatures by the concretization of
the type constructor[s] of the instance, so if all the TyCons match
between a function call and an instance method signature, then we
substitute.

There are quirks involving contexts and forall to pair the arguments
and type constructors up properly for this comparison, and it's still
a work in progress. Notably, the non-recursive transform
`sub_sa_types_T` doesn't account for partial functions by comparing
directly against the saturated type signature. Therefore, partial
applications will fail, which could prematurely stop the pat_match
iteration.