Make patterns located, expand ConPatOut
Analyzing multiple modules now means that we need some way to identify
the bindings that come from the entry point module. The most reliable
way to do that without too much extra language is just locating the
pattern. Note some `HsBind`s don't have locations for the names that
they're binding, but we just need to be in the right file.

ConPatOut also received a case for `InfixCon` which was erroring on
list constructor patterns (e.g. `a:_`).