Clarify `Maybe` semantics in pat_match
Before, there was an awkward transition from the strong signal of
`Nothing` from `pat_match_one` to an empty map in `pat_match_many`.
Thinking a bit about this, we originally returned `Nothing` on failure
so that we could reject MatchGroup alternatives that don't match the
symbols on hand. Furthermore, eagerness isn't necessary from a write-
correctness standpoint: i.e. it's impossible for a failed match to
be hiding a write that could make it succeed if we did evaluate it,
because that violates causality. However, with "scattered" bindings
(the like within `let` and `do`), they wait on writes in general, and
it's far more likely in any case that they fail because of an unknown
function, so we obviously aren't going to fail the whole block because
one of these bindings fails.

Therefore, we propagate the Maybe through `pat_match_many` and expose
a combinator to distinguish between "and" and "or": the former for
arguments, and the latter for scattered bindings.

Also, there is a correction to HsLam handling in `reduce_deep` to
actually make good on that "don't propagate if the arguments fail"
rule. This was lost during the transition to `pat_match_many` a few
commits ago.