`sub` into stack
Top level `pat_match` was failing writes when the write wasn't saturated
in its scope (i.e. it was about to be applied by a function like `$`).
It happened because the binding that needed substitution was in the
stack of the top-level bind that attempted the write, as that binding
wasn't a top-level one by itself.

This doesn't hang at the `everythingWithContextBut` because the
stack isn't itself recursive by construction. More testing may be
needed to confirm that this doesn't cause either of the iterations in
reduce or pat_match to run away infinitely.