Log all `do` statements
Logging only saturated write statements as we were before wasn't general
enough, because they failed the case that the write depended on a symbol
in the same scope, and was therefore unsaturated until pat_match ran.

However, this is still not broad enough, because now bindings in `do`
don't log as writes; the bind matcher can't tell if it's a statement or
not. To fix in the upcoming commits.