Log statements from binds
It turns out the `do` handler is able to log the statements on the right
side of binds by itself by adding statement handling logic to the `BindStmt`
handler which was otherwise just `id`. Now a write that's nested in a
function call with a LHS dependent on a term in the same scope (see `baz`
and `bar` in `J.hs`) can log writes.