Propagate var substitution stack frames
Anti-cycle analysis wasn't working properly: cases akin to `fix` were
failing. This was the initial attempt to solve it.

> A recursive cycle could come from a reference arbitrarily deep in
nested closures, so looking for uniqueness in the topmost string of
`VarRef`s on the stack is not complete enough and would hang. On the other
hand, I've since reasoned that the way stack propagation rules are written
will only record the path taken to get to the binding of a given variable,
and if it's replaced, that stack disappears. Therefore, my initial concern
for looking higher up the stack was unfounded -- I initially thought that
multiple harmless calls to a function in the same stack frame (e.g.
`a \`b\` (c \`b\` d)`) would prevent anything but the first instance from
being applied. In the example above, the inner `b` would be substituted
and the SymApp `c \`b\` d` would be given a VarRefFrame of `b`. However
the outermost `b` call only checks the stack for any `b` in the process
of creating the total expression, and since there aren't any, it too
substitutes for `b` properly.

This explanation was written before I gave it this topic some serious
thought. The example I provided was contrived and worked simply because
the expression was functionally "flat", so the inner and outer `b` calls
didn't need to interact. Things look very different for, say,
`(\x -> x $ ()) $ f` where the inner lambda evaluation will always have
a `$` VarRefFrame in the stack: hence, the VarRefFrame system is actually
inherently flawed. We build towards a correct system in the coming commits.