Limit pat_match iterant cycles and give closures tables
These two approaches together seem to finally resolve `letrec`-based
variable referencing problems. Yet to be addressed fully are function
recursions. This commit addresses things like `let x = f x`. As promised,
here's an elaboration of the problems with the VarRefFrame approach.

> To review the goal: in `let x = if foo then bar else x` I would want to
only extract the possible binding `x = bar`.

> However, it's not as simple as evaluating until I get `x = x`. Pattern
matching breaks this mould. Consider `fix tail` and the definition of `fix f
= let x = f x in x`. In an effort to find out if we reach `let x = x`, we'll
substitute into an infinite `tail (tail (tail...))` because the pattern
match in tail's definition `tail _:t = t` demands eagerness in the leftmost
`:` constructor of its argument. This is contrived for the sake of
simplicity, but by extension, legitimate halting functions with pattern
matches and recursive values won't halt in general in this analysis without
some care.

> On the other hand, it's too hasty to give up on any binding with the same
variable in the left and righthand side, because `(x, y) = (42, x)` is fine.

> It seems that avoiding substituting `x` will work if its result is bound to
be assigned to `x` after pattern matching. However, this doesn't seem to be
straightforward. In `fix tail` above, it forces us to propagate the
information that `x` as an argument becomes bound to `_:t`, and `t` goes on
to be bound to `x`. The trouble is that locally, `_:t <- x` fails the
pattern match.

The point is that anything different than the above condition is going to
fail in type I or type II somehow. I thought up two different ways of
approaching this:

1. Lift a failed pattern match to a first-class citizen of the language and
shuttle it around to see if a variable in the rhshits a pattern of the same
variable. However, this is exceedingly complicated, because we also have to
come up with pattern merging rules, essentially requiring a whole extra set
of language rules.
2. Recognize that the number of terms on each side is finite and forms a
graph of references. The longest paths/cycles in this graph represent the
maximum amount of time it would take a genuine value to reach the farthest
binding before the rest of the behavior is repetitive. Therefore, doing a
simple graph search over these variable bindings will yield a sharp upper
bound for the number of cycles needed to flow the values everywhere.

Not surprisingly, the second option won out. Note there is an accompanying
requirement that I've assumed via educated guess: that the symbols that
depend on things outside of the clique will be resolved in exactly one
pat_match iteration/reduce_deep call, because they're all available in the
stack already. This needs formal verification, but I don't see a hole in it
yet.

---

To a somewhat tangential problem that was tackled simultaneously: I was
horrified to discover that I hadn't ever thought about the fact that
closures can export local bindings, violating my assumption that non-
argument references would always be resolved by a scope above the region
that it's forced via `sub` -- this is only true of zeroth-order values.
Lambdas do need access to the bindings from which they were spawned
(and above, although this is automatically handled by the recursion).
Also, these lambdas can call each other, so that means the stack entry
needed to itself be recursive -- knot-tied so that when one refers to the
other, it has references to the other variables in the clique, which
themselves have access to these variables, so reference chains can work
properly outside of their birthing ground.

This is the only recursive structure (besides the types) and as such it
of course caused problems initially with some of the SYB. After qualifying
everything with a "don't touch the stack" policy, including a rollback of
a recent edit (fc56799), many of the old tests were back up and running.