Check lambda locations to anti-cycle lambdas from pipes
I realized a corner case of the analysis recently while thinking about
binding anti-cycles: one of the fundamental assumptions was that lambdas
don't have to worry about recursion, because the only way for them to get
access to themselves is if they refer to a name that they're bound to.
This is true even of something like `fix` because within the function it
binds its arguments in the argument and `let` clauses.

The exception is IORef and friends, because any scope can grab a reference
to the lambda and shove it into the pipe without the lambda needing an
explicit reference to a variable that it obviously binds to (which would
otherwise be caught by our new pat_match scheme). Instead, this would cause
spinning on the `reduce` iterant for which there is no upper bound, just a
stop condition.

However, it should suffice to just ensure that we don't call a lambda from
the same location more than once (note: callsite is the unique factor, not
definition site).