Merge stacks at the site of substitution in reduce_deep
When lambdas in a letrec have a loop in the call graph, reduce_deep wasn't
previously able to catch it because the substitution would inherit the
stack of the incoming symbol completely. In the simplest case of
`let g x = g x in g ()`, every time `g` was substituted for `\x -> g x`,
the stack leading up to `let` would eliminate the AppFrame containing the
outer call `g ()`, and reduce_deep wouldn't terminate.

However, the stacks to arrive at symbols are important for unique
identification for pipes and consumers. Since substitution doesn't change
the identity of the symbol, it's necessary to also preserve the initial
stack that constructed the symbol, and keep the original stack for lookups
and anti-cycling.

For the moment, the entire stack of the symbol being substituted is appended
to the incoming symbol, which is inefficient and will make debugging more
difficult. I believe that it only suffices to append the frames starting
from the BindFrame itself. However, the order of this is difficult, since
for this to work, the appended frame would need to sit at the head of the
new stack, but this risks intercepting bindings with the same names as
symbols in the scope of the incoming symbol. Some further thought is needed
to improve this solution.