Use lazy `everythingWithContextBut`
The original `everythingWithContextBut` strayed from the SYB definitions,
mapping first (to get the Maybe for decision time), then recursing into
the returned structure. This ends up creating infinite recursion in the
pat_match iterant in `fix`-like cases because the variable is substituted
and then the handler is applied to the _substitution_ which causes the
infinite loop.