Revert reduce_deep call removal
Revert a TODO change from d901849, because the stack can't be in NF in
general. `pat_match` is able to reduce arguments to NF, but it's too
expensive to then reduce all symbols in the stack afterwards, rather
than on a case-by-case basis (upon usage in reduce_deep in this case).