Label SymApps coming from statements
Special-casing `do` and monadic operators for IO-like monads requires
that the SymApps that come from there are labelled, to identify the site
of "execution" for an IO expression. Pipes which are generated by
functions will be incorrectly referenced going off of just the location/
stack in code because individual IO expressions aren't referentially-
transparent and be executed multiple times corresponding to
instantiation of multiple variables.