Propagate symbol location for statements
Lookups continued to be ambiguous because one extra level of indirection
(through AbsBinds in the case of the tests in this case) would prevent
the incoming variable from being substituted with the final binding
location, and once again collapse references to the same term, creating
ambiguity by collapsing distinct pipes into the same one.

Instead, here we also propagate the location into the incoming symbol
even if it isn't a pipe, so that it can eventually propagate into a pipe
as necessary. This raises concerns by changing SymApp locations re: lambda
anti-cycle, but while This might cause some extra iterations, it shouldn't
cause infinite recursion because there are still only finitely many sites
in the program, and lambda anti-cycle is very broad at the moment, only
requiring the function be called from the same site at some point.