Confine location replacement to "executed" expressions
Previously, all variables bound in a `do` block and placed in argument
positions had their locations replaced. The assumption prior was that
only the left-most expression in a binding (e.g. the `foo` in
`v <- foo bar`) would have `StmtFrame` as the first frame, but this
ended up being generally inconsistent with the truth: neither necessary
nor sufficient. A more accurate property that has succeeded basic
tests is that only the topmost syms in a PatMatchSyms (i.e. non-
embedded syms, which for the moment only include those in stacks and
arguments) with a StmtFrame in front of any AppFrame has its location
replaced. Ideally this would correspond only to expressions found in
locations marked with `*` in the following:

1. `do { v <- *; *; * foo }`
2. `* >> *`
3. `* >>= (* -> ...)`

Limited cases of 1. have been tested in `T.hs` and `U.hs`. The
standalone monadic operators remain untested so far.