Expand reads in stacks
This commit provides functionality for the first time to expand the
reads within stack frames so they can create new symbol bindings via
pat_match (this used to be implemented with `Hold`s a while ago). It
became necessary to distinguish BindFrame from AppFrame during HsLam
handling, because the knot-tied table in BindFrames forces the generic
read-expander in `reduce` to hang. Note also I've reaffirmed the
distinction between `pat_match_many` (non-letrec) and `pat_match`
(letrec) behaviors of argument pattern matching and free pattern
matching respectively in `refresh_frames`.

One more note: it's necessary to make the pat_match iterant "lazy"
(terminology to change: "lazy" just means to recurse into a Data,
then apply a transformation on it, as opposed to the other way around)
because it needs to see the changes to its stack before trying to
refresh the frames.