DFS pat_match instead of BFS
Another problem unique to pipes has twisted my hand to make structural
changes. Previously, the BFS for resolving letrec symbols in `pat_match`
sufficed for pure code. However, `pat_match` also handles all the
non-write substitutions within `do` statements, which relocate some
symbols (notably pipe constructors). For pipe resolution to work, these
relocations have to be consistent everywhere they're referenced. The
issue with BFS is that as symbols propagated, some saturated lambdas were
resolved later than they would've been bound by a `<-` in a `do` block.
A slightly more involved read-write then produced a bug:
`do { let mkvar _ = newEmptyMVar; v <- mkvar (); readMVar v }` when `v`
was resolved in `readMVar v` directly to `mkvar ()` from the symbol table.
Then, the next iteration resolved `mkvar` and relocated the returned
`newEmptyMVar` to the callsite in the `readMVar` argument, which was
inconsistent from the `v <- ` binding callsite, which resolved
independently. Then sometimes (for yet unknown reasons why it ever worked)
it wouldn't resolve the pipe because of a location mismatch.

DFS lets us use the execution stack to repeat processing of bindings
exactly. Note the recursion in the new iterant (`dig`) is almost
stateless, only recording the number of iterations to stay below
`max_iters`. Therefore, when it re-encounters a SymApp, regardless of
origin, it will be processed identically modulo recursive variables.
Therefore, the `v` reference in the argument of `readMVar` in the MVE
above would be located to the `v <- ` binding as it should.

It seems that this has caused a regression in half the results, which
is much less catastrophic than it sounds: it seems that one part of the
ABEBinds are no longer resolving at some point. Still, this points to
a location mismatch somewhere else.