Recursively expand reads
Values travelling into and out of many pipes weren't previously captured
because expand_reads was only one level deep, reads that happened to fire
before nested reads had a chance to be resolved were left with an
unresolved read. This might have been a hangover from the previous
"progressive" iteration of `reduce`, which would have resolved these reads
once they bubbled through the chain of writes. However, now that we're
starting from `syms0` again every time, the unresolved read is never
resolved.

Supporting recursive read expansion is not trivial, because it is actually
analogous to `sub` and `pat_match`, where there are potentially recursive
references. In this case, it's even more likely as something like a swap
binds the two variables involved into a recursive pair, since we don't
deliberately don't account for the order or scope of reads and writes.
The recursion is trimmed by an upper bound, in this case the total number
of pipes involved. This is a crude upper bound, but graph analysis for
read expansion is slightly more awkward than pattern matching, because
pattern matching has explicit names and bindings, but this is all
implicit in Binds.

One more change is making the generic recursion explicit for `reduce`'s
`iterant`. This is due to conflicting requirements in the order of
recursion vs. transformation between `refresh_frames`, which expects
the recursion to modify the stack first, vs. recursive read expansion
which needs to transform first without letting the value itself be
`gmapQT`'d in the cases it gives up and returns an unresolved read.