Repair pat_match max_iter graph analysis
I hadn't realized I had made a mistake in the graph analysis for `max_iter`
until now, where I would accumulate all variables in a given `Bind` and
treat them as both pattern and expression variables. This was due to an
accidental use of `&&&` over `***` and was masked to a type error by the
generic nature of `get_ids`, and masked from a behavioral error from the
fact that this is a looser upper bound than the truth. The new routine
must also filter expression vars by whatever is actually bound in the
clique.